pyproject.toml
[tool.poetry]
name = "topping-bot"
version = "0.1.0"
description = ""
authors = ["Will Humphlett <will@humphlett.net>"]
readme = "README.md"
packages = [{include = "topping_bot"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
discord-py = "^2.1.0"
python-dotenv = "^0.21.1"
opencv-python = "^4.7.0.68"
pyyaml = "^6.0"
tqdm = "^4.64.1"
pytz = "^2022.7.1"
tzdata = "^2022.7"
requests = "^2.28.2"
lxml = "^4.9.2"
cssselect = "^1.2.0"
pyparsing = "^3.0.9"
pillow = "^9.5.0"
seaborn = "^0.12.2"
scipy = "^1.10.1"

[tool.poetry.group.dev.dependencies]
psutil = "^5.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
reqconvert = 'topping_bot.util.scripts:req_convert'
assetdump = 'topping_bot.util.scripts:cookie_dump'
